Create Click Logging API with API Key Validation
Related to #699
